Move credit mining module to Core
