+++ Tribler/Core/CreditMining/defs.py
